change according to MVVM architecture
