+++ app/src/main/java/com/example/allholidayscalendar/domain/Interactor.kt
@@ -1,0 +1,10 @@
+package com.example.allholidayscalendar.domain
+
+import com.example.allholidayscalendar.Countries
+import com.example.allholidayscalendar.data.MainRepository
+
+//нужен для того, чтобы забирать нашу базу из слоя model
+class Interactor(val repo: MainRepository) {
+    fun getCountries(): List<Countries> = repo.databaseCountries
+
+}
