+++ app/src/main/java/com/example/allholidayscalendar/view/rv_viewholders/CountriesViewHolder.kt
@@ -1,6 +1,7 @@
-package com.example.allholidayscalendar
+package com.example.allholidayscalendar.view.rv_viewholders
 
 import androidx.recyclerview.widget.RecyclerView
+import com.example.allholidayscalendar.Countries
 import com.example.allholidayscalendar.databinding.CountryItemBinding
 
 class CountriesViewHolder(binding:CountryItemBinding): RecyclerView.ViewHolder(binding.root) {
@@ -10,7 +11,7 @@ class CountriesViewHolder(binding:CountryItemBinding): RecyclerView.ViewHolder(b
     private val poster = binding.poster
 
     //Помещаем данные из Countries в View
-    fun bind(country:Countries){
+    fun bind(country: Countries){
         name.text = country.name
         poster.setImageResource(country.poster)
     }
