+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/master_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
 
     <include layout="@layout/b_scene"/>
