+++ app/src/main/res/layout/fragment_choose.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ChooseFragment"
+    tools:context=".view.fragments.ChooseFragment"
     android:id="@+id/container">
 
     <androidx.appcompat.widget.SearchView
